fastapi/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
pgvector
asyncpg
aiokafka
redis
slowapi
prometheus-fastapi-instrumentator
python-multipart
passlib[bcrypt]
python-jose[cryptography]
